LayoutTests/platform/wpe/imported/w3c/web-platform-tests/XMLHttpRequest/access-control-basic-get-fail-non-simple-expected.txt

CONSOLE MESSAGE: line 22: Preflight response is not successful
CONSOLE MESSAGE: line 22: XMLHttpRequest cannot load http://127.0.0.1:8800/XMLHttpRequest/resources/access-control-basic-options-not-supported.py due to access control checks.

PASS Preflighted cross-origin request denied 
